Bug 654787 - part3: Implementation of audio seamless looping; r?jwwang
To loop audio seamlessly, we keep demuxing, decoding and playing audio in
decoding state while looping. When `HandleEndOfAudio()` is fired, we reset the
demuxer and restart demuxing and decoding data from the beginning, instead of
moving to seeking state then going to decoding state again.

The decoded data is continuously pushed into the audio queue, so the data for
the begining will be placed after the data for the end. To calculate the left
duration in one queue or how many data we left to play till the end, we keep
tracking how many data is pushed or popped and use them to estimate the value.

In original implementation, the looping is implemented by restarting the whole
process to play an audio file with seeking the playback to the beginning.
The `seeking` and `seeked` events will be fired then. However, in this change,
we forece MediaFormatReader to seek in advance when `HandleEndOfAudio()` is fired,
so we never enter the seeking state. Therefore, we also need to fire the `seeking`
and `seeked` events by ourselves in decoding state to meet the HTML spec.

MozReview-Commit-ID: DvdH7yvWbOm