Bug 1523920 - P2: Evaluate SpeechSynthesis state by queue and current task. r?eeejay
It's easier to reuse the utterance if the utterance is stateless. In
original implementation, the state of the utterance is used to ascertain
the state of the SpeechSynthesis. However, the state of the
SpeechSynthesis could be evaluated by the current task and the length of
queue, without using the state of the utterance. This change removes the
concept of the state in the SpeechSynthesisUtterance and make it
reusable.