Bug 1654462 - Import mp4parse c api to get alpha item
This imports the change of https://github.com/mozilla/mp4parse-rust/pull/258

Now the git is refered to my personal git repo. Once the above PR is merged
in the upstream, the git will be updated back to the upstream instead.

This patch might be split to another bug once the PR 258 is merged. For
now, this is introduced to build the D98950.

Differential Revision: https://phabricator.services.mozilla.com/D98950