WIP: Bug 1749539 - Implement basic VideoFrame interfaces
Implement the VideoFrame's Constructor for RGBA ArrayBuffer* and the
basic getters.

This patch pass the wpt test `Test invalid buffer constructed
VideoFrame` by luck since constructing a VideoFrame from ArrayBuffer* is
acceptable now. However, we only support RGBA format in this patch so
that *pass*ed test doesn't be verified correctly. We will address this
in the following patches.

Depends on D145395

Differential Revision: https://phabricator.services.mozilla.com/D146752