Bug 1774300 - Add a blank VideoColorSpace interface
This patch creates a blank class for the VideoColorSpace interface. The
files are generated by running `./mach build-backend && ./mach
webidl-example VideoColorSpace` with necessary changes to make it
buildable.

The VideoColorSpace is a sub-interface of the VideoFrame interface,
which is the essential building block for W3C WebCodecs API.

The implementations are plain blank now. They will be filled out in the
following patches.

Additionally, this patch creates a `dom.media.webcodecs.enabled` pref
for W3C Webcodecs API. All the WebCodecs APIs will be hidden without
setting it to `true`.

Differential Revision: https://phabricator.services.mozilla.com/D144771