Bug 1774300 - Add a blank VideoFrame interface
This patch creates a blank class for the VideoFrame interface. The files
are generated by running `./mach build-backend && ./mach webidl-example
VideoFrame` with necessary changes to make it buildable.

The VideoFrame interface is the essential interface for W3C WebCodecs
API, used to represent the decoded video data, decoded image, and the
data ready to be encoded.

The implementations are plain blank now. They will be filled out in the
following patches.

Depends on D144771

Differential Revision: https://phabricator.services.mozilla.com/D144772