Bug 1864008 - Move libvpx to libgkcodecs
This patch moves libvpx from libgkmedias to libgkcodecs.

When `ac_add_options --with-system-libvpx` is set in mozconfig
(MOZ_SYSTEM_LIBVPX) case, the libvpx installed on the system will be
used and media/libvpx won't be built. In that case, the libvpx APIs
shouldn't be exposed via gkcodecs' symbols.

Since libpixman needs android_getCpuFeatures and _getCpuFamily APIs,
which were built by libvpx previously, cpu-features.c now is required to
be included in its moz.build to get the API it needs.

Differential Revision: https://phabricator.services.mozilla.com/D193599