Bug 1749047 - Add an EncoderTemplate. r?chunmin
It's a lot of code but there is no complexity. Messages are refptrs on this side
because it simplifies the code, remove lots of moves and the messages are const
anyways so uniqueptr don't bring anything.

There are _two_ methods for encoding, because they don't have the same
signature: the video side can take options. Because I didn't want to refactor
yet, the audio side passes in empty options for now.

I'll reflect the changes bug 1865376 in a further patch, but they are similarly
needed.

Depends on D195707

Differential Revision: https://phabricator.services.mozilla.com/D195708