Bug 1749047 - Properly handle reconfiguration and configuration changes. r?chunmin
When we a regular configure call happens, either it's the first configure call
after a reset (or at all), or we have the guarantee that the previous decoder
has been flushed, and so there are no packets in flight.

When we allow the FFmpegVideoEncoder to reconfigure on the fly, we should keep
track of what config each output should use.

Additionally, it's important to destroy / recreate the new EncoderAgent after
the packets have been output, so that ordering is correct. For this, we can
simply dispatch to the current thread, so that the runnable is after all the
output runnables. This was triggered by the WPT
reconfiguring-encoder.https.any.worker.html?vp9_p0, that doesn't output _any_
packet for the 16 first encode message, and the outputs 16 packets when the
encoder is flushed while reconfiguring.

Depends on D196211

Differential Revision: https://phabricator.services.mozilla.com/D196212