Bug 1873051 - Unblock ControlMessage processing if flush failed
Blocking ControlMessage processing until its dispatched task is done
when the flush fails is unnecessary. Instead, the processing should be
unblocked once it dispatches the error handing task when the flush
fails, which means it finishes processing the current ControlMessage.

The blocking was somehow introduced in bug 1865376 [1, 2]. This patch
just restores the behavior it used to be.

[1] https://searchfox.org/mozilla-central/diff/c61b15bde1371bbfda53eb6a5f033b76dc1b4acf/dom/media/webcodecs/DecoderTemplate.cpp#758
[2] https://hg.mozilla.org/mozilla-central/rev/506426f7bfe23c504fc417e1250df92a278aa8cd