Bug 1839993 - Remove mandatory low-latency mode for WebCodecs decoders
Previously, to address the issue of FFmpegDataDecoder always ignoring
errors during the draining process, WebCodecs decoders were forced into
low-latency mode as a workaround to ensure that expected errors for
corrupt/invalid data were generated in WPTs. With upcoming changes to
ensure that the `Drain()` API of decoders return errors when its
underlying API fails unexpectedly, this patch eliminates this compulsion
and set the low-latency mode only when explicitly requested by the user.

Differential Revision: https://phabricator.services.mozilla.com/D206447