Bug 1839993 - Allow Reset() to reject promises during Flush()'s callback
Previously, Reset() lacked the ability to reject promises of the pending
flush requests during the output callback responsible for delivering the
result request (e.g., [1]). This limitation arose because once the
promise was transferred from the FlushMessage to the task delivering the
flush's result, there was no external access to that flush's promise
outside of the dispatched task.

This patch addresses this issue by implementing *pending flush promises*
as outlined in the WebCodecs spec, albeit with a slight modification,
allowing access to any stored pending promise as long as their
associated ID is provided. Conseqently, Reset() gains the capability to
reject all pending promises, while tasks with the corresponding IDs can
retrieve the promises to resolve or reject them based on their
requirements.

[1] https://searchfox.org/mozilla-central/rev/6b39381c0056c0dff0820fffebedf9980adc8883/testing/web-platform/tests/webcodecs/videoDecoder-codec-specific.https.any.js#542

Differential Revision: https://phabricator.services.mozilla.com/D206934